created new svg folder, added IconPencil to it and moved logo into it
